+++ src/fobi/contrib/plugins/form_elements/fields/radio/forms.py
@@ -1,6 +1,6 @@
 __title__ = 'fobi.contrib.plugins.form_elements.fields.select.forms'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
+__copyright__ = '2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('RadioInputForm',)
 
@@ -8,6 +8,7 @@ from django import forms
 from django.utils.translation import ugettext_lazy as _
 
 from fobi.base import BaseFormFieldPluginForm, get_theme
+from fobi.helpers import validate_initial_for_choices
 
 theme = get_theme(request=None, as_instance=True)
 
@@ -70,3 +71,20 @@ class RadioInputForm(forms.Form, BaseFormFieldPluginForm):
         required = False,
         widget = forms.widgets.CheckboxInput(attrs={'class': theme.form_element_checkbox_html_class})
         )
+
+    def clean_initial(self):
+        """
+        Validating the initial value.
+        """
+        return validate_initial_for_choices(self, 'choices', 'initial')
+#
+#        availalble_choices = dict(
+#            get_select_field_choices(self.cleaned_data['choices'])
+#            ).values()
+#
+#        if not self.cleaned_data['initial'] in availalble_choices:
+#            raise forms.ValidationError(
+#                _("Invalid value for initial! Should be any of the "
+#                  "following: {0}".format(','.join(availalble_choices)))
+#                )
+#        return self.cleaned_data['initial']